requirements.txt
adium-theme-ubuntu==0.3.4
alembic==0.8.10
appdirs==1.4.0
arrow==0.10.0
asn1crypto==0.22.0
astroid==1.5.3
audioread==1.0.3
autoenv==1.0.0
backports.functools-lru-cache==1.4
bcrypt==3.1.3
beautifulsoup4==4.4.1
blinker==1.3
cffi==1.10.0
chardet==2.3.0
click==6.7
configobj==5.0.6
configparser==3.5.0
cryptography==2.0.2
dominate==2.3.1
ecdsa==0.13
enum34==1.1.6
Faker==0.7.7
Flask==0.10.1
flask-appconfig==0.11.1
Flask-Assets==0.12
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
Flask-Debug==0.4.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
flask-nav==0.6
Flask-RQ==0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WTF==0.11
fontypython==0.4.4
future==0.16.0
gunicorn==19.6.0
honcho==0.7.1
html5lib==0.999
idna==2.5
inflection==0.3.1
ipaddr==2.1.11
ipaddress==1.0.16
isort==4.2.15
itsdangerous==0.24
Jinja2==2.7.3
jsmin==2.2.1
jsonpickle==0.9.2
jwcrypto==0.4.1
lazy-object-proxy==1.3.1
libvirt-python==1.3.1
lxml==3.5.0
Mako==1.0.6
MarkupSafe==0.23
mccabe==0.6.1
mutagen==1.31
mysql-connector-python-rf==2.2.2
mysqlclient==1.3.7
ndg-httpsclient==0.4.0
numpy==1.11.0
packaging==16.8
passlib==1.7.1
Pillow==3.1.2
psycopg2==2.7
puddletag==1.0.5
pyacoustid==1.1.0
pyasn1==0.1.9
PyBluez==0.18
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pygobject==3.20.0
pylint==1.7.2
pymad==0.9
pymongo==2.8
PyMySQL==0.7.11
pyOpenSSL==17.2.0
pyparsing==2.1.10
python-apt==1.1.0b1
python-dateutil==2.6.0
python-editor==1.0.3
python-jose==1.3.2
raygun4py==3.1.4
redis==2.10.5
requests==2.9.1
rq==0.7.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.5
unity-lens-photos==1.0
urllib3==1.13.1
virtualenv==15.1.0
visitor==0.1.3
webassets==0.12.1
Werkzeug==0.10
wrapt==1.10.10
WTForms==2.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
